+++ libs/cardano-ledger-test/bench/Bench/Cardano/Ledger/StakeDistr.hs
@@ -289,7 +289,7 @@ shelleyGenesis =
       \    \"maxTxSize\": 16384,\
       \    \"maxBlockBodySize\": 65536,\
       \    \"maxBlockHeaderSize\": 1100,\
-      \    \"minFeeA\": 44,\
+      \    \"minFeeFactor\": 44,\
       \    \"minFeeB\": 155381,\
       \    \"minUTxOValue\": 1000000,\
       \    \"poolDeposit\": 500000000,\