+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Conway/Gov.hs
@@ -468,7 +468,7 @@ wfPParamsUpdateSpec =
   constrained' $ \ppupdate ->
     -- Note that ppupdate :: SimplePPUpdate
     match ppupdate $
-      \_minFeeA
+      \_minFeeFactor
        _minFeeB
        maxBBSize
        maxTxSize